refactor folder name for SelectMenu
